Refactor utilFcns to more specific modules
