+++ tools/RAiDER/checkArgs.py
@@ -16,7 +16,7 @@ from datetime import datetime
 
 from RAiDER.constants import Zenith
 from RAiDER.llreader import readLL
-from RAiDER.utilFcns import makeDelayFileNames, modelName2Module
+from RAiDER.ioFcns import makeDelayFileNames
 
 
 def checkArgs(args, p):
@@ -178,3 +178,22 @@ def checkArgs(args, p):
 
     return outArgs
     #return los, lat, lon, bounds, heights, flag, weathers, wmLoc, zref, outformat, datetimeList, out, download_only, verbose, wetNames, hydroNames, parallel
+
+
+def modelName2Module(model_name):
+    """Turn an arbitrary string into a module name.
+    Takes as input a model name, which hopefully looks like ERA-I, and
+    converts it to a module name, which will look like erai. I doesn't
+    always produce a valid module name, but that's not the goal. The
+    goal is just to handle common cases.
+    Inputs:
+       model_name  - Name of an allowed weather model (e.g., 'era-5')
+    Outputs:
+       module_name - Name of the module
+       wmObject    - callable, weather model object
+    """
+    module_name = 'RAiDER.models.' + model_name.lower().replace('-', '')
+    model_module = importlib.import_module(module_name)
+    wmObject = getattr(model_module, model_name.upper().replace('-', ''))
+    return module_name, wmObject
+
