+++ tools/RAiDER/ioFcns.py
@@ -1,4 +1,4 @@
-"""Geodesy-related utility functions."""
+""" General function related to file I/O """
 import importlib
 import multiprocessing as mp
 import os
@@ -16,40 +16,6 @@ from RAiDER import Geo2rdr
 from RAiDER.logger import *
 
 gdal.UseExceptions()
-
-
-def sind(x):
-    """Return the sine of x when x is in degrees."""
-    return np.sin(np.radians(x))
-
-
-def cosd(x):
-    """Return the cosine of x when x is in degrees."""
-    return np.cos(np.radians(x))
-
-
-def lla2ecef(lat, lon, height):
-    ecef = pyproj.Proj(proj='geocent')
-    lla = pyproj.Proj(proj='latlong')
-
-    return pyproj.transform(lla, ecef, lon, lat, height, always_xy=True)
-
-
-def enu2ecef(east, north, up, lat0, lon0, h0):
-    """Return ecef from enu coordinates."""
-    # I'm looking at
-    # https://github.com/scivision/pymap3d/blob/master/pymap3d/__init__.py
-    x0, y0, z0 = lla2ecef(lat0, lon0, h0)
-
-    t = cosd(lat0) * up - sind(lat0) * north
-    w = sind(lat0) * up + cosd(lat0) * north
-
-    u = cosd(lon0) * t - sind(lon0) * east
-    v = sind(lon0) * t + cosd(lon0) * east
-
-    my_ecef = np.stack((x0 + u, y0 + v, z0 + w))
-
-    return my_ecef
 
 
 def gdal_extents(fname):
@@ -167,97 +133,6 @@ def writeArrayToFile(lats, lons, array, filename, noDataValue=-9999):
             f.write('{},{},{}\n'.format(l, L, a))
 
 
-def round_date(date, precision):
-    # First try rounding up
-    # Timedelta since the beginning of time
-    datedelta = datetime.min - date
-    # Round that timedelta to the specified precision
-    rem = datedelta % precision
-    # Add back to get date rounded up
-    round_up = date + rem
-
-    # Next try rounding down
-    datedelta = date - datetime.min
-    rem = datedelta % precision
-    round_down = date - rem
-
-    # It's not the most efficient to calculate both and then choose, but
-    # it's clear, and performance isn't critical here.
-    up_diff = round_up - date
-    down_diff = date - round_down
-
-    return round_up if up_diff < down_diff else round_down
-
-
-def _least_nonzero(a):
-    """Fill in a flat array with the first non-nan value in the last dimension.
-
-    Useful for interpolation below the bottom of the weather model.
-    """
-    mgrid_index = tuple(slice(None, d) for d in a.shape[:-1])
-    return a[tuple(np.mgrid[mgrid_index]) + ((~np.isnan(a)).argmax(-1),)]
-
-
-def robmin(a):
-    '''
-    Get the minimum of an array, accounting for empty lists
-    '''
-    try:
-        return np.nanmin(a)
-    except ValueError:
-        return 'N/A'
-
-
-def robmax(a):
-    '''
-    Get the minimum of an array, accounting for empty lists
-    '''
-    try:
-        return np.nanmax(a)
-    except ValueError:
-        return 'N/A'
-
-
-def _get_g_ll(lats):
-    '''
-    Compute the variation in gravity constant with latitude
-    '''
-    # TODO: verify these constants. In particular why is the reference g different from self._g0?
-    return 9.80616 * (1 - 0.002637 * cosd(2 * lats) + 0.0000059 * (cosd(2 * lats))**2)
-
-
-def _get_Re(lats):
-    '''
-    Returns the ellipsoid as a fcn of latitude
-    '''
-    # TODO: verify constants, add to base class constants?
-    Rmax = 6378137
-    Rmin = 6356752
-    return np.sqrt(1 / (((cosd(lats)**2) / Rmax**2) + ((sind(lats)**2) / Rmin**2)))
-
-
-def _geo_to_ht(lats, hts, g0=9.80556):
-    """Convert geopotential height to altitude."""
-    # Convert geopotential to geometric height. This comes straight from
-    # TRAIN
-    # Map of g with latitude (I'm skeptical of this equation - Ray)
-    g_ll = _get_g_ll(lats)
-    Re = _get_Re(lats)
-
-    # Calculate Geometric Height, h
-    h = (hts * Re) / (g_ll / g0 * Re - hts)
-
-    return h
-
-
-def padLower(invar):
-    '''
-    add a layer of data below the lowest current z-level at height zmin
-    '''
-    new_var = _least_nonzero(invar)
-    return np.concatenate((new_var[:, :, np.newaxis], invar), axis=2)
-
-
 def makeDelayFileNames(time, los, outformat, weather_model_name, out):
     '''
     return names for the wet and hydrostatic delays.
@@ -305,62 +180,6 @@ def make_weather_model_filename(name, time, ll_bounds):
     )
 
 
-def checkShapes(los, lats, lons, hts):
-    '''
-    Make sure that by the time the code reaches here, we have a
-    consistent set of line-of-sight and position data.
-    '''
-    if los is None:
-        los = Zenith
-    test1 = hts.shape == lats.shape == lons.shape
-    try:
-        test2 = los.shape[:-1] == hts.shape
-    except AttributeError:
-        test2 = los is Zenith
-
-    if not test1 and test2:
-        raise ValueError(
-            'I need lats, lons, heights, and los to all be the same shape. ' +
-            'lats had shape {}, lons had shape {}, '.format(lats.shape, lons.shape) +
-            'heights had shape {}, and los was not Zenith'.format(hts.shape))
-
-
-def checkLOS(los, Npts):
-    '''
-    Check that los is either:
-       (1) Zenith,
-       (2) a set of scalar values of the same size as the number
-           of points, which represent the projection value), or
-       (3) a set of vectors, same number as the number of points.
-     '''
-    # los is a bunch of vectors or Zenith
-    if los is not Zenith:
-        los = los.reshape(-1, 3)
-
-    if los is not Zenith and los.shape[0] != Npts:
-        raise RuntimeError('Found {} line-of-sight values and only {} points'
-                           .format(los.shape[0], Npts))
-    return los
-
-
-def modelName2Module(model_name):
-    """Turn an arbitrary string into a module name.
-    Takes as input a model name, which hopefully looks like ERA-I, and
-    converts it to a module name, which will look like erai. I doesn't
-    always produce a valid module name, but that's not the goal. The
-    goal is just to handle common cases.
-    Inputs:
-       model_name  - Name of an allowed weather model (e.g., 'era-5')
-    Outputs:
-       module_name - Name of the module
-       wmObject    - callable, weather model object
-    """
-    module_name = 'RAiDER.models.' + model_name.lower().replace('-', '')
-    model_module = importlib.import_module(module_name)
-    wmObject = getattr(model_module, model_name.upper().replace('-', ''))
-    return module_name, wmObject
-
-
 def read_hgt_file(filename):
     '''
     Read height data from a comma-delimited file
@@ -368,18 +187,6 @@ def read_hgt_file(filename):
     data = pd.read_csv(filename)
     hgts = data['Hgt_m'].values
     return hgts
-
-
-def roundTime(dt, roundTo=60):
-   '''
-   Round a datetime object to any time lapse in seconds
-   dt: datetime.datetime object
-   roundTo: Closest number of seconds to round to, default 1 minute.
-   Source: https://stackoverflow.com/questions/3463930/how-to-round-the-minute-of-a-datetime-object/10854034#10854034
-   '''
-   seconds  = (dt.replace(tzinfo=None) - dt.min).seconds
-   rounding = (seconds+roundTo/2) // roundTo * roundTo
-   return dt + timedelta(0,rounding-seconds,-dt.microsecond)
 
 
 def writeDelays(flag, wetDelay, hydroDelay, lats, lons,
