+++ tools/RAiDER/mathFcns.py
@@ -1,0 +1,110 @@
+""" General math utility functions """
+import importlib
+import multiprocessing as mp
+import os
+import re
+from datetime import datetime, timedelta
+
+import h5py
+import numpy as np
+import pandas as pd
+import pyproj
+from osgeo import gdal, osr
+
+from RAiDER.constants import Zenith
+from RAiDER import Geo2rdr
+from RAiDER.logger import *
+
+gdal.UseExceptions()
+
+
+def sind(x):
+    """Return the sine of x when x is in degrees."""
+    return np.sin(np.radians(x))
+
+
+def cosd(x):
+    """Return the cosine of x when x is in degrees."""
+    return np.cos(np.radians(x))
+
+
+        return data, proj, gt
+
+
+def round_date(date, precision):
+    # First try rounding up
+    # Timedelta since the beginning of time
+    datedelta = datetime.min - date
+    # Round that timedelta to the specified precision
+    rem = datedelta % precision
+    # Add back to get date rounded up
+    round_up = date + rem
+
+    # Next try rounding down
+    datedelta = date - datetime.min
+    rem = datedelta % precision
+    round_down = date - rem
+
+    # It's not the most efficient to calculate both and then choose, but
+    # it's clear, and performance isn't critical here.
+    up_diff = round_up - date
+    down_diff = date - round_down
+
+    return round_up if up_diff < down_diff else round_down
+
+
+def _least_nonzero(a):
+    """Fill in a flat array with the first non-nan value in the last dimension.
+
+    Useful for interpolation below the bottom of the weather model.
+    """
+    mgrid_index = tuple(slice(None, d) for d in a.shape[:-1])
+    return a[tuple(np.mgrid[mgrid_index]) + ((~np.isnan(a)).argmax(-1),)]
+
+
+def robmin(a):
+    '''
+    Get the minimum of an array, accounting for empty lists
+    '''
+    try:
+        return np.nanmin(a)
+    except ValueError:
+        return 'N/A'
+
+
+def robmax(a):
+    '''
+    Get the minimum of an array, accounting for empty lists
+    '''
+    try:
+        return np.nanmax(a)
+    except ValueError:
+        return 'N/A'
+
+
+def padLower(invar):
+    '''
+    add a layer of data below the lowest current z-level at height zmin
+    '''
+    new_var = _least_nonzero(invar)
+    return np.concatenate((new_var[:, :, np.newaxis], invar), axis=2)
+
+
+def roundTime(dt, roundTo=60):
+   '''
+   Round a datetime object to any time lapse in seconds
+   dt: datetime.datetime object
+   roundTo: Closest number of seconds to round to, default 1 minute.
+   Source: https://stackoverflow.com/questions/3463930/how-to-round-the-minute-of-a-datetime-object/10854034#10854034
+   '''
+   seconds  = (dt.replace(tzinfo=None) - dt.min).seconds
+   rounding = (seconds+roundTo/2) // roundTo * roundTo
+   return dt + timedelta(0,rounding-seconds,-dt.microsecond)
+
+
+    else:
+        writeArrayToRaster(wetDelay, wetFilename, noDataValue=ndv,
+                           fmt=outformat, proj=proj, gt=gt)
+        writeArrayToRaster(hydroDelay, hydroFilename, noDataValue=ndv,
+                           fmt=outformat, proj=proj, gt=gt)
+
