+++ tools/RAiDER/processWM.py
@@ -15,7 +15,7 @@ import numpy as np
 from datetime import datetime, date
 
 from RAiDER.logger import *
-from RAiDER.utilFcns import getTimeFromFile
+from RAiDER.ioFcns import getTimeFromFile
 
 
 def getWMFilename(weather_model_name, time, outLoc):
